add -add to server
